refactor: split App.tsx into focused components (1073 → 350 lines)
Extract three components from the App.tsx monolith:

- components/ui/KpiCard.tsx       – renewable/price KPI display card
- components/AppHeader.tsx        – header with live pulse, title, glass buttons
- components/ChartSection.tsx     – chart rendering (renewable, price, scatter)
  including KPI row, ChartDetailView wrappers, view toggle, legends

App.tsx now only orchestrates: state, hooks, contexts, derived data
(hourlyEnergyData, filteredEnergyData, metrics, alertState), and layout.

All 258 tests pass. No behaviour changes.

Part of #301
